karton_plugins/karton-retdec-decompiler/requirements.txt
mwdblib>=4.0.0
karton-core>=4.2.0,<5.0.0
